Big change
Foliage would not render on pre existing landscape. Had to create landscape from scratch.
Decided to delete the level to get a clean slate and get rid of world partition.
Grouped every asset in the level and copy/Pasted it into the new level and formed the landscape to fit.
Made slight adjustment to material.